Team_1/requirements.txt
requests
psutil==5.9.4
python-dotenv
